// lib/api/models/RoleResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RoleResponse = {
    name: string;
    display_name: string;
    system_prompt: string;
    elevenlabs_voice_id: string;
    did_avatar_id: string;
    id: number;
};