// lib/api/models/RoleUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RoleUpdate = {
    name?: (string | null);
    display_name?: (string | null);
    system_prompt?: (string | null);
    elevenlabs_voice_id?: (string | null);
    did_avatar_id?: (string | null);
};